Add macros input in core options depending on the game/system
FBA handles some macros input for certain games/systems.
Street Fighter like games: 3x Punch and 3x Kick
Neogeo: AB, BC, CD ...
PGM: Button 12, 23, 123 ...

It's now possible to choose the macro you want to activate via the core
options.
For Street Fighter only the L2, L3, R2 and R3 buttons will be available.
For the others all the L and R buttons are free to use.
By default, all the macros are set on 'None'.
When you choose a button for a macro, this macro will be activated and
will also appear in the 'Core Input Remapping Options'
The macro core options are affected by game to be able to configure
different macro for each game.