PGM - single pcb boards, missing bios issue
Currently, the following roms will not run with fbalpha (romset verified with clrmame and fba-dat)
...
dmnfrntpcb
svgpcb
thegladpcb
...

fba will just go back to menu when these roms are loaded because its not able to load the bios files from "pgm". the dat files are correct and Mame is able to run the games fine. Discussions from official page does not say anything but just to move the bios files into the rom

-http://neosource.1emulation.com/forums/index.php?topic=1731.645
-http://neosource.1emulation.com/forums/index.php?topic=2981.0

this PR should correct this small issue.

NOTE:
dmnfrntpcb - will show "VERSIONS ARE NOT MATCHED" but it can be bypassed by inserting coins.